+++ src/clip2push_function.ps1
@@ -169,7 +169,7 @@ function clip2push {
         if ($True) {
             if ( -not $Quiet ){
                 [string] $dispName =  (Get-Item -LiteralPath $pushdir).Name
-                Write-Host -Message "in $dispName" -ForegroundColor Green
+                Write-Host "in $dispName" -ForegroundColor Green
             }
             Push-Location -LiteralPath "$pushDir"
             if ( $Pop -And $Action ){